Ruby Exercises

+++ 4.rb
@@ -1,4 +1,4 @@
 puts ("Enter the number")
-   a= gets.chomp.to_i
+   a= gets.chomp.to_f
    b=Math.sqrt(a)
    puts(b)
